EuroToken I: Tokkie (EuroToken's Tikkie) (#1)
EuroToken Group I has made a Tikkie-alike functionality where the user of the superapp can request a Tokkie. The requester can request money in either Euros or EuroTokens to someone, the requested can then pay with both Euros and EuroTokens (excluding Euros to Euros transactions, you can use Tikkie for that). The swapping between Euros and EuroTokens is done through the EuroToken Gateway.