detoks: Add `BasicBlock` class for benchmarking
Add the `BasicBlock` class that we will use later to benchmark the
creation of basic blocks. More specifically, we aim to benchmark
block creation of basic blocks without using IPv8 at all.

For this purpose, we created the `BasicBlock` class that contains all
the necessary information of a basic block. We will use the `BasicBlock`
class for benchmarks that entail "in-memory" storage of blocks or no
storage at all.